Allow multiple connections

+++ source/gdb_server.h
@@ -11,8 +11,8 @@
 typedef struct gdb_server gdb_server_t;
 
 gdb_server_t* gdb_server_create(int port);
-ssize_t gdb_server_waiters(gdb_server_t* server, Waiter* waiters, size_t max);
-bool gdb_server_handle_event(gdb_server_t* server, s32 idx);
+int gdb_server_waiters(gdb_server_t* server, Waiter* waiters, size_t max);
+bool gdb_server_handle_event(gdb_server_t* server, int idx);
 void gdb_server_destroy(gdb_server_t* server);
